Cherrypick some changes from autoscheduler-plugins branch
- move MachineParams from AutoSchedule.h to PIpeline.h
- move some helper functions from AutoSchedule.h to AutoScheduleUtils.h
- minor compilation fix in HalideBuffer.h
