+++ src/Pipeline.cpp
@@ -161,8 +161,8 @@ Pipeline::Pipeline(const vector<Func> &outputs)
 
 vector<Func> Pipeline::outputs() const {
     vector<Func> funcs;
-    for (Function f : contents->outputs) {
-        funcs.push_back(Func(f));
+    for (const Function& f : contents->outputs) {
+        funcs.emplace_back(f);
     }
     return funcs;
 }
@@ -1308,6 +1308,29 @@ JITExtern::JITExtern(Func func)
 
 JITExtern::JITExtern(const ExternCFunction &extern_c_function)
     : extern_c_function_(extern_c_function) {
+}
+
+MachineParams MachineParams::generic() {
+    std::string params = Internal::get_env_variable("HL_MACHINE_PARAMS");
+    if (params.empty()) {
+        return MachineParams(16, 16 * 1024 * 1024, 40);
+    } else {
+        return MachineParams(params);
+    }
+}
+
+std::string MachineParams::to_string() const {
+    std::ostringstream o;
+    o << parallelism << "," << last_level_cache_size << "," << balance;
+    return o.str();
+}
+
+MachineParams::MachineParams(const std::string &s) {
+    std::vector<std::string> v = Internal::split_string(s, ",");
+    user_assert(v.size() == 3) << "Unable to parse MachineParams: " << s;
+    parallelism = std::atoi(v[0].c_str());
+    last_level_cache_size = std::atoll(v[1].c_str());
+    balance = std::atof(v[2].c_str());
 }
 
 }  // namespace Halide
