+++ src/Pipeline.h
@@ -10,7 +10,6 @@
 #include <map>
 #include <vector>
 
-#include "AutoSchedule.h"
 #include "ExternalCode.h"
 #include "IROperator.h"
 #include "IntrusivePtr.h"
@@ -26,6 +25,31 @@ namespace Halide {
 struct Argument;
 class Func;
 struct PipelineContents;
+
+/** A struct representing the machine parameters to generate the auto-scheduled
+ * code for. */
+struct MachineParams {
+    /** Maximum level of parallelism avalaible. */
+    int parallelism;
+    /** Size of the last-level cache (in bytes). */
+    uint64_t last_level_cache_size;
+    /** Indicates how much more expensive is the cost of a load compared to
+     * the cost of an arithmetic operation at last level cache. */
+    float balance;
+
+    explicit MachineParams(int parallelism, uint64_t llc, float balance)
+            : parallelism(parallelism), last_level_cache_size(llc), balance(balance) {
+    }
+
+    /** Default machine parameters for generic CPU architecture. */
+    static MachineParams generic();
+
+    /** Convert the MachineParams into canonical string form. */
+    std::string to_string() const;
+
+    /** Reconstruct a MachineParams from canonical string form. */
+    explicit MachineParams(const std::string &s);
+};
 
 namespace Internal {
 class IRMutator;
